ML: rateLimiter.c - changed the type of "leaked" variable to "unsigned int"
git-svn-id: https://repository.napa-wine.eu/svn/napawine-software/trunk/NAPA-BASELIBS@2379 c54fac2e-1815-407e-9b96-088d06743913

+++ ml/util/rateLimiter.c
@@ -114,7 +114,7 @@ int outputRateControl(int len) {
 		bib_then = now;
 		return OK;
 	} else {
-		long leaked;
+		unsigned int leaked;
 		int total_drain_secs = bytes_in_bucket / (drain_rate) + 1;
 
 		if(now.tv_sec - bib_then.tv_sec - 1 > total_drain_secs) {
